fix(앱): 판가 센트 절삭 + 지연/주의 개수가 0 으로 뜨던 문제
1) 앱이 판가·재료비를 .toInt() 로 잘랐다. $0.7 짜리 부품이 0 이 되면서
   재료비율이 0.0% 로 떴다. double 로 받고 Fmt.unit() 으로 센트까지 쓴다
   ($4,950 / $0.70). Admin 입력·저장은 원래 소수점 둘째 자리까지 받고 있었다.

2) 카드의 지연/주의 개수는 손으로 적는 status 만 보고, 목록 줄은
   완료예정일로 따졌다. 파워박스 목록에 '지연중' 이 다섯인데 카드는 0 이었다.
   서버에 _model_alert() 를 두고 한 번만 정한다.
     손으로 적은 '지연'/'주의' > 완료예정일 지남(지연) > 7일 이내(주의)
   진행률 100% 는 정상. 양산은 완료예정일이 없어 날짜로 따지지 않는다.
   앱은 서버가 준 alert 를 세고, 없으면 옛 status 로 떨어진다.

+++ mobile/lib/screens/model_cost_list_screen.dart
@@ -100,8 +100,10 @@ class _ModelCostListScreenState extends State<ModelCostListScreen> {
                         itemCount: filtered.length,
                         itemBuilder: (context, i) {
                           final m = filtered[i];
-                          final price = (m['price'] as num?)?.toInt() ?? 0;
-                          final mcost = (m['material_cost'] as num?)?.toInt() ?? 0;
+                          // .toInt() 로 자르면 \$0.7 부품이 0 이 되고
+                          // 재료비율이 0.0% 로 뜬다. 센트를 살린다.
+                          final price = (m['price'] as num?)?.toDouble() ?? 0;
+                          final mcost = (m['material_cost'] as num?)?.toDouble() ?? 0;
                           final ratio = price > 0 ? (mcost / price * 100) : null;
                           final group = m['group'] ?? '양산';
                           return GestureDetector(
